Tách database Firestore ra 1 file riêng
Tách ra để mọi người mốt có viết query để get hoặc set data thì chỉ cần tạo file mới trong folder include rồi import {database} from "./Firebase" là ok.

Có thể coi code mẫu trong "include/MenuFecth.js"

Note: viết query tạo file js trong folder include chứ đừng tạo ngoài để cho nó gọn
